feat: tenth commit
Implemented new Stage Selection screen thanks to Samuel_X7 and finally learnt
how to cause certain robot masters to be skipped based on when they can
be accessed or not thru a dictionary of variables by causing the selection
Number to be subtracted or added by 1 when current one is not available.
also working on Wily Stage 2.