Add SlumberMan boss and improve RogueRiot logic
Refactorde RogueRiot boss logic for attack selection esp. to prevent attacks happenign 3x in a row, hitbox activation, and sound effects.
Enhances enemy hurt flash effect, updates audio bus layout to allow for consistent volume by the use of Compressor and Limiter filters (which are dope ngl since they limit and balance loud and quiet BG respectively), and improves menu volume handling by the use of linear_to_db in AudioBus code in BGM main node as well as making the volume from 0 to -60dB to -1.0 to 0 in ratio form for linear_to_db and the AudioBus to work with.
Minor scene and script adjustments for better gameplay and UI consistency.
Also added SlumberMan boss assets and cutscene animation (which is a silly gag xd)

We thank God  🙏 ✨