72nd: Bacckup B4 Updating to Godot 4.5: Add Flameman boss ,logging system,recreating Megaman with state engine
--BACKUP BEFORE UPDATING ENGINE TO GODOT 4.5--
Introduced Flameman boss with state scripts, scene, and assets.
Added a Logger utility for improved logging. Updated various scripts to use Logger and made minor code cleanups, asset additions, and scene adjustments for camera and enemy spawners.
Recreating Megaman ,with a State engine on a different node in order to easily create custom attacks and animations to Mega man + to easily add characters if need be.

Glory Be To GOD 🙏 ✨